flight_deal_finder/requirements.txt
python-dotenv>=1.0.1
httpx>=0.24.1
twilio>=9.1.1
tenacity>=8.2.2
pandas>=2.1.0
python-dateutil>=2.9.2